pyproject.toml
[project]
name = "team6-usecase4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.7.4",
    "autogen-ext[azure,openai]>=0.7.4",
    "azure-identity>=1.24.0",
    "dotenv>=0.9.9",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
]
